all good and almost perfect

+++ database/migrations/2024_07_30_040723_create_users_table.php
@@ -11,7 +11,7 @@ class CreateUsersTable extends Migration
     Schema::create('users', function (Blueprint $table) {
       $table->id();
       $table->string('name');
-      $table->unsignedBigInteger('nrp')->unique();
+      $table->unsignedInteger('nrp')->unique();
       $table->string('email')->unique();
       $table->string('password')->nullable();
       $table->enum('role', ['user', 'admin', 'technician']);
